1.5.85

+++ webapp/Scripts/Security.RoleMember.js
@@ -6,10 +6,10 @@ app.SecurityRoleMember = (function () {
         var filter = $('#RoleMemberSearch').val();
         var recordStatusFltVal = $('#RecordStatusFlt').select2('data').map(function (item) { return item['id']; }).toString();
         if (recordStatusFltVal != '')
-            recordStatusFltVal = '&recordStatus=' + recordStatusFltVal;
+            recordStatusFltVal = '|recordStatus.in.(' + recordStatusFltVal + ")";
 
         $('#RoleMemberGridTbl').bootstrapTable('showLoading');
-        app.core.Get(app.setting.apipath + 'v1/RoleMember/Get?filter=' + filter + recordStatusFltVal)
+        app.core.Get(app.setting.apipath + 'v1/RoleMember?filter=' + filter + recordStatusFltVal)
             .done(function (data, textStatus, jqXHR) {
                 $('#RoleMemberGridTbl').bootstrapTable('load', data !== null ? data : []);
             }).always(function () {
@@ -39,6 +39,7 @@ app.SecurityRoleMember = (function () {
             pagination: true,
             smartDisplay: true,
             detailView: false,
+            showColumns: false,
             detailFormatter: 'app.ui.GenericDetailFormatter',
             columns: [
                 {
@@ -47,7 +48,9 @@ app.SecurityRoleMember = (function () {
                     sortable: true,
                     halign: 'center',
                     events: 'RoleMemberGridTbl_Events',
-                    formatter: 'app.ui.EditLinkFormatter'
+                    formatter: function (value, row, index, field) {
+                        return '<a class="edit" href="javascript:void(0)" title="Al hacer click permite la edición de los datos del rol">' + value + '</a>';
+                    }
                 }, {
                     field: 'Description',
                     title: 'Descripción',
@@ -56,10 +59,17 @@ app.SecurityRoleMember = (function () {
                 }, {
                     field: 'SecurityLevel',
                     title: 'Seguridad',
+                    titleTooltip: 'Nivel de seguridad.',					
                     sortable: true,
                     halign: 'center',
                     align: 'right',
                     formatter: 'app.ui.IntegerFormatter'
+                }, {
+                    field: 'InitialNavigationCodeDesc',
+                    title: 'Página inicial',
+                    sortable: true,
+                    halign: 'center',
+                    formatter: 'app.ui.StringFormatter'
                 }, {
                     field: 'RecordStatusDesc',
                     title: 'Estado',
@@ -77,7 +87,7 @@ app.SecurityRoleMember = (function () {
                 }, {
                     title: 'Acciones',
                     class: 'd-none d-sm-table-cell',
-                    titleTooltip: 'Acciones disponibles para una fila',
+                    titleTooltip: 'Acciones disponibles para un rol',
                     sortable: false,
                     halign: 'center',
                     align: 'center',
@@ -85,8 +95,8 @@ app.SecurityRoleMember = (function () {
                     widthUnit: "%",
                     events: 'RoleMemberGridTbl_Events',
                     formatter: function (value, row, index, field) {
-                        return '<button type="button" class="btn btn-sm btn-white edit" title="Al hacer click permite la edición de la fila"> <i class="fa fa-pencil"></i> </button>' +
-                            '<button type="button" class="btn btn-sm btn-white delete" title="Al hacer click permite eliminar una fila"> <i class="fa fa-close"></i> </button>';
+                        return '<button type="button" class="btn btn-sm btn-white edit" title="Al hacer click permite la edición de los datos del rol de la fila"> <i class="fa fa-pencil"></i> </button>' +
+                            '<button type="button" class="btn btn-sm btn-white delete" title="Al hacer click permite eliminar los datos del rol de la fila"> <i class="fa fa-close"></i> </button>';
                     },
                     cellStyle: function (value, row, index) {
                         return {
@@ -130,7 +140,7 @@ app.SecurityRoleMember = (function () {
         $('#RoleMemberEdtFormSave').click(function () {
             if (app.ui.IsValid('#RoleMemberEdtForm', false)) {
                 app.ui.ButtonDoing('#RoleMemberEdtFormSave');
-                var data = MapInputToObject()
+                var data = MapInputToObject();
                 if (data.RoleId === 0)
                     Create(data, 'Save');
                 else
@@ -142,41 +152,40 @@ app.SecurityRoleMember = (function () {
         $('#RoleMemberEdtFormSaveContinue').click(function () {
             if (app.ui.IsValid('#RoleMemberEdtForm', false)) {
                 app.ui.ButtonDoing('#RoleMemberEdtFormSaveContinue');
-                var data = MapInputToObject()
-                if (data.RoleId === 0)
-                    Create(data, 'Continue');
-                else
-                    Update(data);
+                var data = MapInputToObject();
+                Create(data, 'Continue');
             }
             event.preventDefault();
         });
+
         $('#RoleMemberEdtFormSaveCopy').click(function () {
             if (app.ui.IsValid('#RoleMemberEdtForm', false)) {
                 app.ui.ButtonDoing('#RoleMemberEdtFormSaveCopy');
-                var data = MapInputToObject()
-                if (data.RoleId === 0)
-                    Create(data, 'Copy');
-                else
-                    Update(data);
+                var data = MapInputToObject();
+                Create(data, 'Copy');
             }
             event.preventDefault();
         });
+
         $('#RoleMemberEdtFormDuplicate').click(function () {
             if (app.ui.IsValid('#RoleMemberEdtForm', false)) {
                 app.ui.ButtonDoing('#RoleMemberEdtFormDuplicate');
                 var data = MapInputToObject();
                 data.RoleId = 0;
-                data.RoleName += ' copia';
+                data.RoleName += ' duplicado';
                 app.ui.ButtonDone('#RoleMemberEdtFormDuplicate');
                 app.SecurityRoleMember.EditRow(data);
             }
             event.preventDefault();
         });
+
         $('#RoleMemberEdtFormDelete').click(function () {
-            app.ui.ButtonDoing('#RoleMemberEdtFormDelete');
-            var data = MapInputToObject();
-            app.ui.ButtonDone('#RoleMemberEdtFormDelete');
-            app.SecurityRoleMember.DeleteRow(data);
+            if (app.ui.IsValid('#RoleMemberEdtForm', false)) {
+                app.ui.ButtonDoing('#RoleMemberEdtFormDelete');
+                var data = MapInputToObject();
+                app.ui.ButtonDone('#RoleMemberEdtFormDelete');
+                app.SecurityRoleMember.DeleteRow(data);
+            }
             event.preventDefault();
         });
 
@@ -188,7 +197,7 @@ app.SecurityRoleMember = (function () {
     }
 
     function Create(uidata, mode) {
-        app.core.Post(app.setting.apipath + 'v1/RoleMember/Post', JSON.stringify(uidata))
+        app.core.Post(app.setting.apipath + 'v1/RoleMember', JSON.stringify(uidata))
             .done(function (data, textStatus, jqXHR) {
                 toastr.success("El rol '" + uidata.RoleName + "' fue creado", "", { timeOut: 5000, closeButton: true, progressBar: true });
                 Refresh();
@@ -220,7 +229,7 @@ app.SecurityRoleMember = (function () {
     }
 
     function Update(uidata) {
-        app.core.Put(app.setting.apipath + 'v1/RoleMember/Put?id=' + uidata.RoleId, JSON.stringify(uidata))
+        app.core.Put(app.setting.apipath + 'v1/RoleMember/' + uidata.RoleId, JSON.stringify(uidata))
             .done(function (data, textStatus, jqXHR) {
                 toastr.success("El rol '" + uidata.RoleName + "' fue modificado", "", { timeOut: 5000, closeButton: true, progressBar: true });
                 ViewMode();
@@ -236,7 +245,7 @@ app.SecurityRoleMember = (function () {
                 timeOut: 5000, closeButton: true, progressBar: true,
                 onclick: function () {
                     $('.ibox-content').toggleClass('sk-loading');
-                    app.core.Delete(app.setting.apipath + 'v1/RoleMember/Delete?id=' + uidata.RoleId)
+                    app.core.Delete(app.setting.apipath + 'v1/RoleMember/' + uidata.RoleId)
                         .done(function (data, textStatus, jqXHR) {
                             toastr.success("El rol '" + uidata.RoleName + "' fue eliminado", "", { timeOut: 5000, closeButton: true, progressBar: true });
                             ViewMode();
@@ -249,7 +258,7 @@ app.SecurityRoleMember = (function () {
     }
 
     function Init_Lookups() {
-        app.core.Lookups(['EstadoRegistro.RecordStatus', 'EstadoRegistro.RecordStatusFlt'], Dynamic_Event_Controls);
+        app.core.Lookups(['NavigacionPage.InitialNavigationCode.', 'EstadoRegistro.RecordStatus', 'EstadoRegistro.RecordStatusFlt'], Dynamic_Event_Controls);
         // Dependencies
     }
 
@@ -259,14 +268,15 @@ app.SecurityRoleMember = (function () {
     }
 
     function MapInputToObject() {
-        var data = {
+        return {
             RoleId: parseInt(0 + $('#RoleId').val(), 10),
             RoleName: $('#RoleName').val(),
             Description: $('#Description').val(),
             SecurityLevel: app.ui.GetNumericValue('#SecurityLevel'),
+            InitialNavigationCode: $('#InitialNavigationCode').val(),
             RecordStatus: $('#RecordStatus').val()
+
         };
-        return data;
     }
 
     function MapObjectToInput(data) {
@@ -274,7 +284,9 @@ app.SecurityRoleMember = (function () {
         $('#RoleName').val(data.RoleName);
         $('#Description').val(data.Description);
         app.ui.SetNumericValue('#SecurityLevel', data.SecurityLevel);
+        $('#InitialNavigationCode').val(data.InitialNavigationCode);
         $('#RecordStatus').val(data.RecordStatus);
+
     }
 
     function Setup_Validations() {
@@ -329,9 +341,10 @@ app.SecurityRoleMember = (function () {
         $('.advancefilter-row').addClass('d-none');
         $('#RoleMemberGridTbl').parents().find('.table-responsive').addClass('d-none');
         $('#RoleMemberEdt').removeClass('d-none');
-        MapObjectToInput(row);
-        $('#RoleName').focus();
+
         if (row.RoleId === 0) {
+            MapObjectToInput(row);
+            $('#RoleName').focus();
             $('#RoleMemberEdtFormSaveContinue').removeClass('d-none');
             $('#RoleMemberEdtFormSaveCopy').removeClass('d-none');
             $('#RoleMemberEdtFormDuplicate').addClass('d-none');
@@ -339,12 +352,20 @@ app.SecurityRoleMember = (function () {
             $('#RoleMemberEdtFormSave').removeClass('d-none');
             $('#RoleMemberEdtFormCancel').removeClass('d-none');
         } else {
-            $('#RoleMemberEdtFormSaveContinue').addClass('d-none');
-            $('#RoleMemberEdtFormSaveCopy').addClass('d-none');
-            $('#RoleMemberEdtFormDuplicate').removeClass('d-none');
-            $('#RoleMemberEdtFormDelete').removeClass('d-none');
-            $('#RoleMemberEdtFormSave').removeClass('d-none');
-            $('#RoleMemberEdtFormCancel').removeClass('d-none');
+            $('.ibox-content').toggleClass('sk-loading');
+            app.core.Get(app.setting.apipath + 'v1/RoleMember/' + row.RoleId)
+                .done(function (data, textStatus, jqXHR) {
+                    MapObjectToInput(data);
+                    $('#RoleName').focus();
+                    $('#RoleMemberEdtFormSaveContinue').addClass('d-none');
+                    $('#RoleMemberEdtFormSaveCopy').addClass('d-none');
+                    $('#RoleMemberEdtFormDuplicate').removeClass('d-none');
+                    $('#RoleMemberEdtFormDelete').removeClass('d-none');
+                    $('#RoleMemberEdtFormSave').removeClass('d-none');
+                    $('#RoleMemberEdtFormCancel').removeClass('d-none');
+                }).always(function () {
+                    $('.ibox-content').toggleClass('sk-loading');
+                });
         }
     }
 
@@ -366,9 +387,9 @@ app.SecurityRoleMember = (function () {
             Setup_Validations();
         },
         New: function (row) {
-            let newRow = { RoleId: 0, RoleName: null, Description: null, SecurityLevel: 50, RecordStatus: 1 }
+            let newRow = { RoleId: 0, RoleName: null, Description: null, SecurityLevel: 50, InitialNavigationCode: null, RecordStatus: 1 }
             if (row !== undefined) {
-                row.Roleid = 0;
+                row.RoleId = 0;
                 newRow = row;
             }
             EditMode(newRow);
